Adding support for darwin to IsVTXDisabled
