+++ drivers/virtualbox/virtualbox.go
@@ -187,24 +187,6 @@ func (d *Driver) PreCreateCheck() error {
 	return d.vbm()
 }
 
-// IsVTXDisabled checks if VT-X is disabled in the BIOS. If it is, the vm will fail to start.
-// If we can't be sure it is disabled, we carry on and will check the vm logs after it's started.
-func (d *Driver) IsVTXDisabled() bool {
-	if runtime.GOOS == "windows" {
-		output, err := cmdOutput("wmic", "cpu", "get", "VirtualizationFirmwareEnabled")
-		if err != nil {
-			log.Debugf("Couldn't check that VT-X/AMD-v is enabled. Will check that the vm is properly created: %v", err)
-			return false
-		}
-
-		disabled := strings.Contains(output, "FALSE")
-		return disabled
-	}
-
-	// TODO: linux and OSX
-	return false
-}
-
 // cmdOutput runs a shell command and returns its output.
 func cmdOutput(name string, args ...string) (string, error) {
 	cmd := exec.Command(name, args...)
