+++ drivers/virtualbox/virtualbox_linux.go
@@ -1,0 +1,7 @@
+package virtualbox
+
+// IsVTXDisabled checks if VT-X is disabled in the BIOS. If it is, the vm will fail to start.
+// If we can't be sure it is disabled, we carry on and will check the vm logs after it's started.
+func (d *Driver) IsVTXDisabled() bool {
+	return false
+}
